requirements.txt
cairocffi
CairoSVG
certifi
cffi
chardet
charset-normalizer
chess
contourpy
cssselect2
cycler
defusedxml
filelock
filetype
fonttools
fsspec
idna
Jinja2
kiwisolver
Levenshtein
lxml
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python
opencv-python-headless
packaging
pandas
pillow
psutil
py-cpuinfo
pycairo
pycparser
pyparsing
python-chess
python-dateutil
python-dotenv
python-Levenshtein
pytz
PyYAML
RapidFuzz
reportlab
requests
requests-toolbelt
roboflow
scipy
seaborn
shapely
six
svglib
sympy
tabulate
tinycss2
torch
torchvision
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
webencodings
